fix(csa-resource): bind extra-readable paths at resolved dest (#3075)
