+++ weave.lock
@@ -1,6 +1,6 @@
 [versions]
-csa = "0.1.1144"
-weave = "0.1.1144"
+csa = "0.1.1145"
+weave = "0.1.1145"
 last_migrated_at = "2026-03-08T12:08:01.820964091Z"
 
 [migrations]
